reused normalization code
