+++ pkg/scheduling/plugins/scorer/prefix_aware.go
@@ -87,50 +87,3 @@ func (s *PrefixAwareScorer) PostSchedule(ctx *types.SchedulingContext, res *type
 func (s *PrefixAwareScorer) GetPrefixStore() *PrefixStore {
 	return s.prefixStore
 }
-
-// podToKey is a function type that converts a Pod to a string key.
-// It returns the key and a boolean indicating success.
-type podToKeyFunc func(pod types.Pod) (string, bool)
-
-func indexedScoresToNormalizedScoredPods(pods []types.Pod, podToKey podToKeyFunc,
-	scores map[string]int) map[types.Pod]float64 {
-	scoredPods := make(map[types.Pod]float64)
-	minScore, maxScore := getMinMax(scores)
-
-	for _, pod := range pods {
-		key, ok := podToKey(pod)
-		if !ok {
-			continue
-		}
-
-		if score, ok := scores[key]; ok {
-			if minScore == maxScore {
-				scoredPods[pod] = 1.0
-				continue
-			}
-
-			scoredPods[pod] = float64(score-minScore) / float64(maxScore-minScore)
-		} else {
-			scoredPods[pod] = 0.0
-		}
-	}
-
-	return scoredPods
-}
-
-// TODO: share with other scorers?
-func getMinMax(scores map[string]int) (int, int) {
-	minScore := int(^uint(0) >> 1) // max int
-	maxScore := -1
-
-	for _, score := range scores {
-		if score < minScore {
-			minScore = score
-		}
-		if score > maxScore {
-			maxScore = score
-		}
-	}
-
-	return minScore, maxScore
-}
